put Install at end
Reasoning: When trying to figure out how to do something I don't know if it's basic or advanced. So I often find myself having to repeatedly leap back and forth over the Install section, and that's friction.

Theortically, Install is referred to once or twice, and then perhaps never again, while looking up this or that usage is a more frequent and repeated action.